refactor(broker): split requestor_monitor.rs into directory layout (C40 of PR-3)
Pure mechanical move. requestor_monitor.rs (880 lines) becomes:

- requestor_monitor/mod.rs (~33 lines): module decls + pub use
  service::RequestorMonitor + pub use types::{AllowRequestors,
  PriorityRequestors}.
- requestor_monitor/error.rs (~30 lines): MonitorError enum +
  coded_error_impl!.
- requestor_monitor/types.rs (~227 lines): CachedEntry (private),
  PriorityRequestors + AllowRequestors structs and their impls.
  Two private methods are widened to pub(super) so service.rs can drive
  them across modules: update_from_list and clear_from_url. The config
  field is widened to pub(super) for the same reason — service.rs
  reads it directly to drive the periodic refresh.
- requestor_monitor/service.rs (~664 lines): RequestorMonitor struct +
  impl, BrokerService impl, the REFRESH_INTERVAL constant, and the
  original test module.

External paths preserved:
- crate::requestor_monitor::PriorityRequestors (broker.rs,
  order_evaluator, utils/storage, order_committer, proving, order_pricer)
- crate::requestor_monitor::AllowRequestors (broker.rs, order_pricer)
- crate::requestor_monitor::RequestorMonitor (broker.rs)
