+++ crates/broker/src/requestor_monitor/mod.rs
@@ -1,0 +1,33 @@
+// Copyright 2026 Boundless Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Requestor monitor — periodically refreshes the priority and allowed
+//! requestor lists from configured remote sources, exposing the results
+//! through [`PriorityRequestors`] and [`AllowRequestors`] handles held by
+//! downstream services.
+//!
+//! Layout:
+//! - [`service`] — the [`RequestorMonitor`] struct, its constructor, and the
+//!   [`BrokerService`](crate::task::BrokerService) `run` loop that drives
+//!   the periodic refresh.
+//! - [`types`] — [`PriorityRequestors`] and [`AllowRequestors`] query
+//!   handles + the internal `CachedEntry` row.
+//! - [`error`] — [`MonitorError`].
+
+mod error;
+mod service;
+mod types;
+
+pub use service::RequestorMonitor;
+pub use types::{AllowRequestors, PriorityRequestors};
