+++ crates/broker/src/requestor_monitor/service.rs
@@ -12,232 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    coded_error_impl,
-    config::ConfigLock,
-    errors::CodedError,
-    task::{BrokerService, SupervisorErr},
-};
-use alloy::primitives::Address;
+use crate::task::{BrokerService, SupervisorErr};
 use anyhow::Result;
-use requestor_lists::{Extensions, RequestorEntry, RequestorList};
-use std::{
-    collections::HashMap,
-    sync::{Arc, RwLock},
-};
-use thiserror::Error;
+use requestor_lists::RequestorList;
 use tokio::time::Duration;
 use tokio_util::sync::CancellationToken;
 
+use super::error::MonitorError;
+use super::types::{AllowRequestors, PriorityRequestors};
+
 const REFRESH_INTERVAL: Duration = Duration::from_secs(3600); // 1 hour
-
-#[derive(Error)]
-pub enum MonitorError {
-    #[error("Failed to lock internal state")]
-    LockFailed,
-}
-
-coded_error_impl!(MonitorError, "RM",
-    LockFailed => "4001",
-);
-
-/// Cached entry that tracks both the requestor data and its source URL
-#[derive(Clone, Debug)]
-struct CachedEntry {
-    entry: RequestorEntry,
-    source_url: String,
-}
-
-/// Tracks priority requestors from both static config and remote lists
-#[derive(Clone, Debug)]
-pub struct PriorityRequestors {
-    /// Map of requestor addresses to their cached entries (from remote lists)
-    requestors: Arc<RwLock<HashMap<Address, CachedEntry>>>,
-    /// Config lock for reading latest static addresses
-    config: ConfigLock,
-    /// Chain ID for this broker instance
-    chain_id: u64,
-}
-
-/// Tracks allowed requestors from both static config and remote lists
-#[derive(Clone, Debug)]
-pub struct AllowRequestors {
-    /// Map of requestor addresses to their cached entries (from remote lists)
-    requestors: Arc<RwLock<HashMap<Address, CachedEntry>>>,
-    /// Config lock for reading latest static addresses
-    config: ConfigLock,
-    /// Chain ID for this broker instance
-    chain_id: u64,
-}
-
-impl PriorityRequestors {
-    pub fn new(config: ConfigLock, chain_id: u64) -> Self {
-        Self { requestors: Arc::new(RwLock::new(HashMap::new())), config, chain_id }
-    }
-
-    /// Get requestor entry for an address, checking both remote lists and static config
-    pub fn get_requestor_entry(&self, address: &Address) -> Option<RequestorEntry> {
-        // First check cached remote list entries
-        if let Ok(requestors) = self.requestors.read() {
-            if let Some(cached) = requestors.get(address) {
-                return Some(cached.entry.clone());
-            }
-        }
-
-        // Then check static config addresses
-        if let Ok(config) = self.config.lock_all() {
-            if let Some(static_addresses) = &config.market.priority_requestor_addresses {
-                if static_addresses.contains(address) {
-                    // Create a default entry for static config addresses (no extensions)
-                    return Some(RequestorEntry {
-                        address: *address,
-                        chain_id: self.chain_id,
-                        name: "Static Config Requestor".to_string(),
-                        description: None,
-                        tags: vec!["static".to_string()],
-                        extensions: Extensions::default(),
-                    });
-                }
-            }
-        }
-
-        None
-    }
-
-    /// Check if an address is a priority requestor
-    pub fn is_priority_requestor(&self, address: &Address) -> bool {
-        self.get_requestor_entry(address).is_some()
-    }
-
-    /// Returns all dynamically-registered priority addresses (from remote lists).
-    pub fn dynamic_addresses(&self) -> Vec<Address> {
-        self.requestors.read().map(|r| r.keys().copied().collect()).unwrap_or_default()
-    }
-
-    /// Update the priority requestors from a list
-    fn update_from_list(&self, list: &RequestorList, source_url: &str) {
-        let mut requestors = match self.requestors.write() {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::error!("Failed to lock requestors for writing: {e:?}");
-                return;
-            }
-        };
-
-        let mut added_count = 0;
-        for entry in &list.requestors {
-            // Only add requestors for the chain we're operating on
-            if entry.chain_id == self.chain_id {
-                requestors.insert(
-                    entry.address,
-                    CachedEntry { entry: entry.clone(), source_url: source_url.to_string() },
-                );
-                added_count += 1;
-            }
-        }
-
-        tracing::info!(
-            "Updated priority requestors from list '{}' (v{}.{}, schema v{}.{}): {} entries added (chain_id={}, {} total in list)",
-            list.name,
-            list.version.major,
-            list.version.minor,
-            list.schema_version.major,
-            list.schema_version.minor,
-            added_count,
-            self.chain_id,
-            list.requestors.len()
-        );
-    }
-
-    /// Clear cached priority requestors from a specific URL
-    fn clear_from_url(&self, url: &str) {
-        if let Ok(mut requestors) = self.requestors.write() {
-            requestors.retain(|_, cached| cached.source_url != url);
-        }
-    }
-}
-
-impl AllowRequestors {
-    pub fn new(config: ConfigLock, chain_id: u64) -> Self {
-        Self { requestors: Arc::new(RwLock::new(HashMap::new())), config, chain_id }
-    }
-
-    /// Get requestor entry for an address, checking both remote lists and static config
-    pub fn get_requestor_entry(&self, address: &Address) -> Option<RequestorEntry> {
-        // First check cached remote list entries
-        if let Ok(requestors) = self.requestors.read() {
-            if let Some(cached) = requestors.get(address) {
-                return Some(cached.entry.clone());
-            }
-        }
-
-        // Then check static config addresses
-        if let Ok(config) = self.config.lock_all() {
-            if let Some(static_addresses) = &config.market.allow_client_addresses {
-                if static_addresses.contains(address) {
-                    // Create a default entry for static config addresses (no extensions)
-                    return Some(RequestorEntry {
-                        address: *address,
-                        chain_id: self.chain_id,
-                        name: "Static Config Requestor".to_string(),
-                        description: None,
-                        tags: vec!["static".to_string()],
-                        extensions: Extensions::default(),
-                    });
-                }
-            }
-        }
-
-        None
-    }
-
-    /// Check if an address is an allowed requestor
-    pub fn is_allow_requestor(&self, address: &Address) -> bool {
-        self.get_requestor_entry(address).is_some()
-    }
-
-    /// Update the allow requestors from a list
-    fn update_from_list(&self, list: &RequestorList, source_url: &str) {
-        let mut requestors = match self.requestors.write() {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::error!("Failed to lock requestors for writing: {e:?}");
-                return;
-            }
-        };
-
-        let mut added_count = 0;
-        for entry in &list.requestors {
-            // Only add requestors for the chain we're operating on
-            if entry.chain_id == self.chain_id {
-                requestors.insert(
-                    entry.address,
-                    CachedEntry { entry: entry.clone(), source_url: source_url.to_string() },
-                );
-                added_count += 1;
-            }
-        }
-
-        tracing::info!(
-            "Updated allowed requestors from list '{}' (v{}.{}, schema v{}.{}): {} entries added (chain_id={}, {} total in list)",
-            list.name,
-            list.version.major,
-            list.version.minor,
-            list.schema_version.major,
-            list.schema_version.minor,
-            added_count,
-            self.chain_id,
-            list.requestors.len()
-        );
-    }
-
-    /// Clear cached allowed requestors from a specific URL
-    fn clear_from_url(&self, url: &str) {
-        if let Ok(mut requestors) = self.requestors.write() {
-            requestors.retain(|_, cached| cached.source_url != url);
-        }
-    }
-}
 
 /// Service for periodically monitoring and refreshing requestor priority and allowed lists
 #[derive(Clone)]
